requirements.txt
# Core dependencies
requests>=2.28.0
PyYAML>=6.0

# Data analysis dependencies for optimization analyzer
pandas>=1.5.0
numpy>=1.21.0

# Optional dependencies for enhanced features
urllib3>=1.26.0

# Database and analytics dependencies
# SQLite3 is built into Python, no additional dependencies needed

# Development dependencies (optional)
pytest>=7.0.0         # Testing framework
pytest-cov>=4.0.0     # Coverage reporting
